fix content-length bug

+++ src/main/java/com/volcengine/tos/RequestOptions.java
@@ -9,7 +9,7 @@ import java.time.format.DateTimeFormatter;
 
 public class RequestOptions {
 
-    public static RequestOptionsBuilder withContentLength(int length) {
+    public static RequestOptionsBuilder withContentLength(long length) {
         return builder -> builder.withContentLength(length);
     }
